data_processing/requirements.txt
# Core data pipeline
duckdb>=0.10.0

# ML pipeline (hra_ml_insights.py)
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
prophet>=1.1
ruptures>=1.1

# GitHub data fetching (fetch_cns_github.py)
pyyaml>=6.0

# Testing
pytest>=8.0
